商品管理　ほしいもの ・javascript処理を全てwebpackでapp.jsにまとまるように修正
