+++ application/public/mix-manifest.json
@@ -1,4 +1,5 @@
 {
+    "/js/admin/app.js": "/js/admin/app.js",
     "/js/app.js": "/js/app.js",
     "/css/app.css": "/css/app.css",
     "/css/admin.css": "/css/admin.css",
