+++ application/resources/js/admin/app.js
@@ -1,0 +1,3 @@
+require('../bootstrap');
+require('../jquery.maskMoney.min');
+require('./products.js');
