+++ application/resources/js/admin/products.js
@@ -1,0 +1,15 @@
+$(function(){
+    $(".form_money").on("submit", function(){
+        $(".input_money").each((index, element) => {
+            element.value = element.value.replace(/[¥,]/g, "")
+        })
+    })
+
+    $(".input_money").maskMoney({
+        prefix:"¥",
+        thousands:",",
+        allowZero:true,
+        precision:"0",
+    })
+        .maskMoney("mask")
+})
